Interface fixação frete

+++ 0x08/resumo_pedido/Pedido.java
@@ -5,7 +5,7 @@ import resumo_pedido.provedores.*;
 public class Pedido {
 
     private int codigo;
-    private int peso;
+    private double peso;
     private double total;
     private Frete frete;
 
@@ -24,7 +24,7 @@ public class Pedido {
         this.codigo = codigo;
     }
 
-    public int getPeso() {
+    public double getPeso() {
         return peso;
     }
